refactor(util): move decorators into irsim/util/decorator.py
plot_only, normalize_actions, time_it and time_it2 leave the 1,200-line util.py for a module of their own. The irsim.util package now exports them together with the random and message helpers, and EnvBase imports from the package, so the layout underneath can change without touching importers.
